frontend: redux/configSlice: Refactor config into a configSlice
