+++ frontend/src/components/App/Layout.tsx
@@ -10,8 +10,8 @@ import { useDispatch } from 'react-redux';
 import { request } from '../../lib/k8s/apiProxy';
 import { Cluster } from '../../lib/k8s/cluster';
 import { getCluster } from '../../lib/util';
-import { setConfig } from '../../redux/actions/actions';
-import { ConfigState } from '../../redux/reducers/config';
+import { setConfig } from '../../redux/configSlice';
+import { ConfigState } from '../../redux/configSlice';
 import { useTypedSelector } from '../../redux/reducers/reducers';
 import store from '../../redux/stores/store';
 import ActionsNotifier from '../common/ActionsNotifier';
