+++ frontend/src/components/App/Notifications/List.stories.tsx
@@ -2,7 +2,7 @@ import { configureStore } from '@reduxjs/toolkit';
 import { Meta, Story } from '@storybook/react/types-6-0';
 import helpers from '../../../helpers';
 import { Notification } from '../../../lib/notification';
-import { INITIAL_STATE as CONFIG_INITIAL_STATE } from '../../../redux/reducers/config';
+import { initialState as CONFIG_INITIAL_STATE } from '../../../redux/configSlice';
 import { INITIAL_STATE as FILTER_INITIAL_STATE } from '../../../redux/reducers/filter';
 import { INITIAL_STATE as UI_INITIAL_STATE } from '../../../redux/reducers/ui';
 import { TestContext } from '../../../test';
