+++ frontend/src/components/App/Settings/Settings.tsx
@@ -4,8 +4,9 @@ import { useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useDispatch } from 'react-redux';
 import LocaleSelect from '../../../i18n/LocaleSelect/LocaleSelect';
-import { setAppSettings, setVersionDialogOpen } from '../../../redux/actions/actions';
-import { defaultTableRowsPerPageOptions } from '../../../redux/reducers/config';
+import { setVersionDialogOpen } from '../../../redux/actions/actions';
+import { setAppSettings } from '../../../redux/configSlice';
+import { defaultTableRowsPerPageOptions } from '../../../redux/configSlice';
 import { ActionButton, NameValueTable, SectionBox } from '../../common';
 import TimezoneSelect from '../../common/TimezoneSelect';
 import { useSettings } from './hook';
