+++ frontend/src/redux/reducers/reducers.tsx
@@ -2,9 +2,9 @@ import { TypedUseSelectorHook, useSelector } from 'react-redux';
 import { combineReducers } from 'redux';
 import pluginsReducer from '../../plugin/pluginsSlice';
 import actionButtons from '../actionButtonsSlice';
+import configReducer from '../configSlice';
 import detailsViewSectionsSlice from '../detailsViewSectionsSlice';
 import clusterAction from './clusterAction';
-import config from './config';
 import filter from './filter';
 import uiReducer from './ui';
 
@@ -12,7 +12,7 @@ const reducers = combineReducers({
   filter: filter,
   ui: uiReducer,
   clusterAction: clusterAction,
-  config: config,
+  config: configReducer,
   plugins: pluginsReducer,
   actionButtons: actionButtons,
   detailsViewSections: detailsViewSectionsSlice,
